requirements.txt
pylint==2.4.4
pdfkit==0.6.1
jira==2.0.0
python-dotenv==0.12.0
dataclasses==0.7